requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.2
Pillow==10.0.0